refactor(routes): move endpoints to blueprints
